pyproject.toml
[project]
name = "api-to-mcp"
version = "0.1.0"
description = "Convert Web APIs to MCP Servers"
authors = [
    {name = "Your Name", email = "your.email@example.com"}
]
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "requests>=2.31.0",
    "openai>=1.0.0",
    "pydantic>=2.0.0",
    "click>=8.1.0",
    "jinja2>=3.1.0",
    "pyyaml>=6.0.0",
    "fastmcp>=2.0.0",
    "httpx>=0.25.0",
    "streamlit>=1.28.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0.0",
    "black>=23.0.0",
    "ruff>=0.1.0",
]

[project.scripts]
api-to-mcp = "api_to_mcp.cli:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/api_to_mcp"]

[tool.black]
line-length = 100
target-version = ['py310']

[tool.ruff]
line-length = 100
target-version = "py310"

